requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
banal==1.0.6
bitsets==0.8.4
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
concepts==0.9.2
contourpy==1.3.0
cycler==0.12.1
dataset==1.6.2
datasets==3.1.0
dill==0.3.8
Django==4.2.17
exceptiongroup==1.2.2
fastapi==0.115.6
features==0.5.12
ffmpy==0.4.0
fileconfig==0.6.1
filelock==3.16.1
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.9.0
gradio==4.44.1
gradio_client==1.3.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
image==1.5.33
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.2.1
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.8.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlparse==0.5.2
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
tomlkit==0.12.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.15
uvicorn==0.32.1
value==0.1.0
websockets==12.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
